add editing customer - incomleted

+++ src/Pages/Customers/Customers.jsx
@@ -5,6 +5,7 @@ import CustomerTable from "../Dashboard/CustomerTable";
 import {AiOutlineUser} from 'react-icons/ai';
 import UserInfoForm from "../Orders.V2/UserInfoForm";
 import useToggle from "../Orders.V2/customHooks/useToggle";
+import { useDeferredValue, useState } from "react";
 function Customers(props) {
     const [customers, setCustomers] = useData({
         store: STORES.CUSTOMERS.name,
@@ -12,9 +13,16 @@ function Customers(props) {
         keyPath: null
     });
 
+    const [query, setQuery] = useState("");
+    const deferredQuery = useDeferredValue(query);
     const [addFormToggle, setAddFormToggle] = useToggle(false);
+    const filteredCustomers = customers.filter((customer) => {
+        return customer.customerName.toLowerCase().includes(deferredQuery.toLowerCase());
+    });
+    const isStale = query !== deferredQuery;
 
     async function deleteCustomer(id) {
+        if (confirm("Are you sure you want to delete this customer?") === false) return;
         await setCustomers({
             type: "delete",
             indexField: STORES.CUSTOMERS.keyPath,
@@ -23,6 +31,9 @@ function Customers(props) {
     }
 
     function updateCustomer(id, newVal) {
+        alert("update customer");
+        console.log(newVal);
+        return;
         setCustomers({
             type: "update",
             indexField: STORES.CUSTOMERS.keyPath,
@@ -57,7 +68,7 @@ function Customers(props) {
                         <button className="btn text-primary " onClick={setAddFormToggle}>Add Customer</button>
                         
                         <div className="input-group " style={{width: "15em"}}>
-                        <input type="text" className="form-control" placeholder="Search" aria-label="Search" aria-describedby="basic-addon2" />
+                        <input type="text" className="form-control" placeholder="Search" aria-label="Search" aria-describedby="basic-addon2" value={query} onChange={(e) => setQuery(e.target.value)}/>
                         <span className="input-group-text" id="basic-addon2">Search</span>
                         </div>
                     </div>
@@ -66,12 +77,18 @@ function Customers(props) {
 
 
 
-            <div className="row">
+            <div className="row" 
+                style={{
+                    opacity: isStale ?"0.5" : "1",
+                    transition: isStale ?  'opacity 0.2s 0.2s linear' : 'opacity 0s 0s linear',
+                }}
+            >
                 <CustomerTable 
-                    sortedCustomers={customers}
+                    sortedCustomers={filteredCustomers}
                     deleteCustomer={deleteCustomer}
                     updateCustomer={updateCustomer}
                 />
+                
             </div>
         </>
      );
